+++ core/src/avm1/object/script_object.rs
@@ -607,6 +607,7 @@ mod tests {
                 video: &mut NullVideoBackend::new(),
                 mouse_over_object: None,
                 mouse_down_object: None,
+                input: &Default::default(),
                 mouse_position: &(Twips::ZERO, Twips::ZERO),
                 drag_object: &mut None,
                 player: None,
