+++ core/src/context.rs
@@ -12,7 +12,7 @@ use crate::backend::{
     navigator::NavigatorBackend,
     render::RenderBackend,
     storage::StorageBackend,
-    ui::UiBackend,
+    ui::{InputManager, UiBackend},
     video::VideoBackend,
 };
 use crate::context_menu::ContextMenuState;
@@ -100,6 +100,9 @@ pub struct UpdateContext<'a, 'gc, 'gc_context> {
 
     /// If the mouse is down, the display object that the mouse is currently pressing.
     pub mouse_down_object: Option<DisplayObject<'gc>>,
+
+    /// The input manager, tracking keys state.
+    pub input: &'a InputManager,
 
     /// The location of the mouse when it was last over the player.
     pub mouse_position: &'a (Twips, Twips),
@@ -297,6 +300,7 @@ impl<'a, 'gc, 'gc_context> UpdateContext<'a, 'gc, 'gc_context> {
             stage: self.stage,
             mouse_over_object: self.mouse_over_object,
             mouse_down_object: self.mouse_down_object,
+            input: self.input,
             mouse_position: self.mouse_position,
             drag_object: self.drag_object,
             player: self.player.clone(),
