+++ core/src/display_object/edit_text.rs
@@ -1344,14 +1344,14 @@ impl<'gc> EditText<'gc> {
                 let length = text.len();
                 match key_code {
                     ButtonKeyCode::Left => {
-                        if (context.ui.is_key_down(KeyCode::Shift) || selection.is_caret())
+                        if (context.input.is_key_down(KeyCode::Shift) || selection.is_caret())
                             && selection.to > 0
                         {
                             selection.to = string_utils::prev_char_boundary(text, selection.to);
-                            if !context.ui.is_key_down(KeyCode::Shift) {
+                            if !context.input.is_key_down(KeyCode::Shift) {
                                 selection.from = selection.to;
                             }
-                        } else if !context.ui.is_key_down(KeyCode::Shift) {
+                        } else if !context.input.is_key_down(KeyCode::Shift) {
                             selection.to = selection.start();
                             selection.from = selection.to;
                         }
@@ -1360,14 +1360,14 @@ impl<'gc> EditText<'gc> {
                         return ClipEventResult::Handled;
                     }
                     ButtonKeyCode::Right => {
-                        if (context.ui.is_key_down(KeyCode::Shift) || selection.is_caret())
+                        if (context.input.is_key_down(KeyCode::Shift) || selection.is_caret())
                             && selection.to < length
                         {
                             selection.to = string_utils::next_char_boundary(text, selection.to);
-                            if !context.ui.is_key_down(KeyCode::Shift) {
+                            if !context.input.is_key_down(KeyCode::Shift) {
                                 selection.from = selection.to;
                             }
-                        } else if !context.ui.is_key_down(KeyCode::Shift) {
+                        } else if !context.input.is_key_down(KeyCode::Shift) {
                             selection.to = selection.end();
                             selection.from = selection.to;
                         }
