+++ core/src/player.rs
@@ -12,7 +12,7 @@ use crate::backend::{
     navigator::{NavigatorBackend, RequestOptions},
     render::RenderBackend,
     storage::StorageBackend,
-    ui::{MouseCursor, UiBackend},
+    ui::{InputManager, MouseCursor, UiBackend},
     video::VideoBackend,
 };
 use crate::config::Letterbox;
@@ -203,6 +203,8 @@ pub struct Player {
     /// Faked time passage for fooling hand-written busy-loop FPS limiters.
     time_offset: u32,
 
+    input: InputManager,
+
     mouse_pos: (Twips, Twips),
     is_mouse_down: bool,
 
@@ -233,7 +235,6 @@ pub struct Player {
     current_frame: Option<u16>,
 }
 
-#[allow(clippy::too_many_arguments)]
 impl Player {
     #[allow(clippy::too_many_arguments)]
     pub fn new(
@@ -295,6 +296,8 @@ impl Player {
             frame_accumulator: 0.0,
             recent_run_frame_timings: VecDeque::with_capacity(10),
             time_offset: 0,
+
+            input: Default::default(),
 
             mouse_pos: (Twips::ZERO, Twips::ZERO),
             is_mouse_down: false,
@@ -839,11 +842,16 @@ impl Player {
     /// 9. Mouse state is updated. This triggers button rollovers, which are a
     ///    second wave of event processing.
     pub fn handle_event(&mut self, event: PlayerEvent) {
+        self.input.handle_event(&event);
+
         if cfg!(feature = "avm_debug") {
             match event {
                 PlayerEvent::KeyDown {
                     key_code: KeyCode::V,
-                } if self.ui.is_key_down(KeyCode::Control) && self.ui.is_key_down(KeyCode::Alt) => {
+                    ..
+                } if self.input.is_key_down(KeyCode::Control)
+                    && self.input.is_key_down(KeyCode::Alt) =>
+                {
                     self.mutate_with_update_context(|context| {
                         let mut dumper = VariableDumper::new("  ");
                         let levels: Vec<_> = context.stage.iter_depth_list().collect();
@@ -874,7 +882,10 @@ impl Player {
                 }
                 PlayerEvent::KeyDown {
                     key_code: KeyCode::D,
-                } if self.ui.is_key_down(KeyCode::Control) && self.ui.is_key_down(KeyCode::Alt) => {
+                    ..
+                } if self.input.is_key_down(KeyCode::Control)
+                    && self.input.is_key_down(KeyCode::Alt) =>
+                {
                     self.mutate_with_update_context(|context| {
                         if context.avm1.show_debug_output() {
                             log::info!(
@@ -905,9 +916,8 @@ impl Player {
                     key_code: ButtonKeyCode::from_u8(codepoint as u8).unwrap(),
                 })
             }
-
             // Special keys have custom values for keyPress.
-            PlayerEvent::KeyDown { key_code } => {
+            PlayerEvent::KeyDown { key_code, .. } => {
                 if let Some(key_code) = crate::events::key_code_to_button_key_code(key_code) {
                     Some(ClipEvent::KeyPress { key_code })
                 } else {
@@ -1512,6 +1522,7 @@ impl Player {
                 stage,
                 mouse_over_object: mouse_hovered_object,
                 mouse_down_object: mouse_pressed_object,
+                input: &self.input,
                 mouse_position: &self.mouse_pos,
                 drag_object,
                 player: self.self_reference.clone(),
